ingestao-kruzer-produtos-pbm/requirements.txt
requests==2.32.3
pandas==2.2.2
pyarrow==17.0.0
google-cloud-storage==2.18.2
dotenv
google-cloud-logging